style: tweak LoginPage
