+++ docs/applications/application_functions.md
