+++ exp/key_value_store/res/kv_store.wasm
