+++ exp/key_value_store_context/res/key_value_store_context.wasm
