+++ exp/mining_pool/build.sh
