+++ exp/mining_pool/res/mining_pool.wasm
