+++ src/app/chat/build.sh
