+++ src/app/chat/res/chat.wasm
