+++ src/app/mining/miner/build.sh
