+++ src/app/mining/miner/res/miner.wasm
