+++ src/backend/external_contract/res/external_contract.wasm
