+++ src/backend/proxy_contract/build.sh
