+++ src/backend/proxy_contract/res/proxy_contract.wasm
